Refactor BridgeFileUploadManager to pull out the sandbox file path methods
