+++ SwiftPackage/Sources/BridgeClientExtension/File Uploading/V1/ParticipantFileUploadAPI.swift
@@ -65,7 +65,7 @@ public class ParticipantFileUploadAPI: BridgeFileUploadAPITyped {
         
         // Set up a directory to keep temp copies of files being uploaded
         do {
-            let appSupportDir = try FileManager.default.sharedUploadDirectory()
+            let appSupportDir = try FileManager.default.sharedAppSupportDirectory()
             self.tempUploadDirURL = appSupportDir.appendingPathComponent("ParticipantFileUploads")
             try FileManager.default.createDirectory(at: self.tempUploadDirURL, withIntermediateDirectories: true, attributes: nil)
         }
